requirements.txt
# ── Core backend ─────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pyyaml==6.0.1
httpx==0.27.0
aiofiles==23.2.1
websockets==12.0
jinja2==3.1.3
python-multipart==0.0.9

# ── Auth ─────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ── Database ─────────────────────────────────────────
sqlalchemy==2.0.27
aiosqlite==0.19.0
alembic==1.13.1

# ── LLM providers ────────────────────────────────────
anthropic>=0.40.0
openai>=1.30.0
google-generativeai>=0.7.0
ollama>=0.2.0

# ── Model Context Protocol ───────────────────────────
mcp>=1.0.0

# ── Tool execution & sandboxing ──────────────────────
ptyprocess==0.7.0
psutil==5.9.8

# ── Screenshots & recording ──────────────────────────
pillow==10.2.0
pyvirtualdisplay==3.0
asciinema==2.4.0

# ── Reporting ────────────────────────────────────────
markdown==3.5.2
weasyprint==61.2
rich==13.7.0

# ── Networking / pentest helpers ─────────────────────
python-nmap==0.7.1
dnspython==2.6.1
requests==2.31.0
beautifulsoup4==4.12.3

# ── Observability ────────────────────────────────────
structlog==24.1.0
loguru==0.7.2

# ── Dev ──────────────────────────────────────────────
pytest==8.0.2
pytest-asyncio==0.23.5
ruff==0.3.0
